+++ src/Parquet.Test/Integration/ParquetMrIntegrationTest.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using Parquet.Data;
-using Parquet.Data.Rows;
+using Parquet.Rows;
 using Parquet.Schema;
 using Xunit;
 using F = System.IO.File;
