+++ src/Parquet.Test/Rows/RowsModelTest.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Parquet.Data;
-using Parquet.Data.Rows;
+using Parquet.Rows;
 using Parquet.Schema;
 using Xunit;
