--- src/Parquet/Data/Rows/StringFormat.cs
@@ -1,9 +1,0 @@
-﻿namespace Parquet.Data.Rows
-{
-   enum StringFormat
-   {
-      JsonSingleQuote,
-      Json,
-      Csv
-   }
-}
