+++ src/Parquet/Rows/StringFormat.cs
@@ -1,0 +1,7 @@
+﻿namespace Parquet.Rows {
+    enum StringFormat {
+        JsonSingleQuote,
+        Json,
+        Csv
+    }
+}
